requirements.txt
coreapi==2.3.1
Django==1.11.4
django-debug-toolbar==1.8
django-filter==1.0.4
djangorestframework==3.6.3
psycopg2==2.7.3
pytz==2017.2
